requirements.txt
# Core dependencies
pandas==2.2.0
numpy==1.26.4
sqlalchemy>=2.0.0
psycopg2-binary
python-dotenv
requests

# Airflow (version 3.1.3)
apache-airflow==3.1.3
apache-airflow-providers-postgres>=5.0.0